allow 1/5/10 increments for long press move developer panel src to /offroad/settings
